Rename and reorganize swift prevalidation DTOs
