--- BisHackIso20022Api/src/main/java/com/bishack/api/dto/ClearingSystemMemberIdentification.java
@@ -1,43 +1,0 @@
-package com.bishack.api.dto;
-
-import java.util.HashMap;
-import java.util.Map;
-import com.fasterxml.jackson.annotation.JsonAnyGetter;
-import com.fasterxml.jackson.annotation.JsonAnySetter;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-"member_identification"
-})
-public class ClearingSystemMemberIdentification {
-
-@JsonProperty("member_identification")
-private String memberIdentification;
-@JsonIgnore
-private Map<String, Object> additionalProperties = new HashMap<String, Object>();
-
-@JsonProperty("member_identification")
-public String getMemberIdentification() {
-return memberIdentification;
-}
-
-@JsonProperty("member_identification")
-public void setMemberIdentification(String memberIdentification) {
-this.memberIdentification = memberIdentification;
-}
-
-@JsonAnyGetter
-public Map<String, Object> getAdditionalProperties() {
-return this.additionalProperties;
-}
-
-@JsonAnySetter
-public void setAdditionalProperty(String name, Object value) {
-this.additionalProperties.put(name, value);
-}
-
-}
