+++ BisHackIso20022Api/src/main/java/com/bishack/api/dto/swift/prevalidation/accountfmt/VerifyAccountFmtReqDto.java
@@ -1,7 +1,8 @@
-package com.bishack.api.dto;
+package com.bishack.api.dto.swift.prevalidation.accountfmt;
 
 import org.springframework.stereotype.Component;
 
+import com.bishack.api.dto.BaseDto;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
@@ -11,15 +12,12 @@ import com.fasterxml.jackson.annotation.JsonRootName;
 @JsonIgnoreProperties(ignoreUnknown = true)
 @Component
 @JsonRootName(value = "SwiftPrevalAcFormatDto")
-public class SwiftPrevalAcFormatDto extends BaseDto {
-
-	/**
-	 * 
-	 */
+public class VerifyAccountFmtReqDto extends BaseDto {
 	private static final long serialVersionUID = 1L;
 	private String account_identification;
 	private String country_code;
 	private String financial_institution_identification;
+
 	public String getAccount_identification() {
 		return account_identification;
 	}
@@ -38,14 +36,14 @@ public class SwiftPrevalAcFormatDto extends BaseDto {
 	public void setFinancial_institution_identification(String financial_institution_identification) {
 		this.financial_institution_identification = financial_institution_identification;
 	}
-	public SwiftPrevalAcFormatDto(String account_identification, String country_code,
+	public VerifyAccountFmtReqDto(String account_identification, String country_code,
 			String financial_institution_identification) {
 		super();
 		this.account_identification = account_identification;
 		this.country_code = country_code;
 		this.financial_institution_identification = financial_institution_identification;
 	}
-	public SwiftPrevalAcFormatDto() {
+	public VerifyAccountFmtReqDto() {
 		super();
 
 	}
