+++ BisHackIso20022Api/src/main/java/com/bishack/api/service/SwiftApiService.java
@@ -15,13 +15,13 @@ import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
 
 import com.bishack.api.dto.SwiftInstitutionReviewDto;
-import com.bishack.api.dto.SwiftPrevalAcFormatDto;
-import com.bishack.api.dto.SwiftPrevalAcVerifyDto;
 import com.bishack.api.dto.SwiftPrevalBicDto;
 import com.bishack.api.dto.SwiftPrevalCatPurposeDto;
 import com.bishack.api.dto.SwiftPrevalPurposeCdDto;
 import com.bishack.api.dto.SwiftPrevalPurposeDto;
 import com.bishack.api.dto.SwiftTokenDto;
+import com.bishack.api.dto.swift.prevalidation.account.VerifyAccountReqDto;
+import com.bishack.api.dto.swift.prevalidation.accountfmt.VerifyAccountFmtReqDto;
 
 @Service(value = "swiftApiService")
 public class SwiftApiService implements ISwiftApiService {
@@ -34,7 +34,7 @@ public class SwiftApiService implements ISwiftApiService {
 	private ISwiftApiTokenService swiftApiTokenService;
 
 	@Override
-	public String swiftPrevalAcFormat(SwiftPrevalAcFormatDto swiftPrevalAcFormatDto) throws Exception {
+	public String swiftPrevalAcFormat(VerifyAccountFmtReqDto swiftPrevalAcFormatDto) throws Exception {
 		String response = null;
 
 		if (swiftPrevalAcFormatDto != null) {
@@ -46,7 +46,7 @@ public class SwiftApiService implements ISwiftApiService {
 				headers.add("Authorization", "Bearer " + swiftTokenDto.getAccess_token());
 				headers.setContentType(MediaType.APPLICATION_JSON);
 
-				HttpEntity<SwiftPrevalAcFormatDto> request = new HttpEntity<>(swiftPrevalAcFormatDto, headers);
+				HttpEntity<VerifyAccountFmtReqDto> request = new HttpEntity<>(swiftPrevalAcFormatDto, headers);
 
 				ResponseEntity<String> responseEnt = swiftApiRestTemplate.exchange(
 						"https://sandbox.swift.com/swift-preval-pilot/v2/payment/account-format", HttpMethod.POST,
@@ -60,7 +60,7 @@ public class SwiftApiService implements ISwiftApiService {
 	}
 
 	@Override
-	public String swiftPrevalAcVerify(SwiftPrevalAcVerifyDto swiftPrevalAcVerifyDto, String bic) throws Exception {
+	public String swiftPrevalAcVerify(VerifyAccountReqDto swiftPrevalAcVerifyDto, String bic) throws Exception {
 		String response = null;
 		
 
@@ -74,7 +74,7 @@ public class SwiftApiService implements ISwiftApiService {
 				headers.setContentType(MediaType.APPLICATION_JSON);
 				headers.add("x-bic", bic);
 
-				HttpEntity<SwiftPrevalAcVerifyDto> request = new HttpEntity<>(swiftPrevalAcVerifyDto, headers);
+				HttpEntity<VerifyAccountReqDto> request = new HttpEntity<>(swiftPrevalAcVerifyDto, headers);
 
 				ResponseEntity<String> responseEnt = swiftApiRestTemplate.exchange(
 						"https://sandbox.swift.com/swift-preval-pilot/v2/accounts/verification", HttpMethod.POST,
